fix bug get wrong session id stories

+++ app/src/main/java/net/iGap/controllers/MessageDataStorage.java
@@ -974,7 +974,7 @@ public class MessageDataStorage extends BaseController {
                             putStoriesToDatabase(database, stories.get(i).getSeenAllGroupStories(), stories.get(i).getOriginatorId(), storyObjects, stories.get(i).getOriginatorName(), realmRegisteredInfo != null ? realmRegisteredInfo.getColor() : "#4aca69");
                             storyObjects.removeAll(storyObjects);
                         } else {
-                            RealmStory realmStory = database.where(RealmStory.class).equalTo("userId", stories.get(i).getOriginatorId()).findFirst();
+                            RealmStory realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", stories.get(i).getOriginatorId()).findFirst();
                             if (realmStory != null) {
                                 realmStory.deleteFromRealm();
                             }
@@ -994,7 +994,7 @@ public class MessageDataStorage extends BaseController {
                             }
                             if (!isExist) {
                                 long userId = realmStories.get(i).getUserId();
-                                RealmStory realmStory = database.where(RealmStory.class).equalTo("userId", userId).findFirst();
+                                RealmStory realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findFirst();
                                 if (realmStory != null && realmStory.isSentAll()) {
                                     realmStory.deleteFromRealm();
                                 }
@@ -1016,7 +1016,7 @@ public class MessageDataStorage extends BaseController {
                             putStoriesToDatabase(database, stories.get(i).getSeenAllGroupStories(), stories.get(i).getOriginatorId(), storyObjects, stories.get(i).getOriginatorName(), realmRegisteredInfo != null ? realmRegisteredInfo.getColor() : "#4aca69");
                             storyObjects.removeAll(storyObjects);
                         } else {
-                            RealmStory realmStory = database.where(RealmStory.class).equalTo("userId", stories.get(i).getOriginatorId()).findFirst();
+                            RealmStory realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", stories.get(i).getOriginatorId()).findFirst();
                             if (realmStory != null) {
                                 realmStory.deleteFromRealm();
                             }
@@ -1030,7 +1030,7 @@ public class MessageDataStorage extends BaseController {
                     if (realmStories != null && realmStories.size() > 0) {
                         for (int i = 0; i < realmStories.size(); i++) {
                             if (realmStories.get(i).getRealmStoryProtos().size() == 0) {
-                                database.where(RealmStory.class).equalTo("userId", realmStories.get(i).getUserId()).findAll().deleteAllFromRealm();
+                                database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", realmStories.get(i).getUserId()).findAll().deleteAllFromRealm();
                             }
                         }
                     }
@@ -1188,7 +1188,7 @@ public class MessageDataStorage extends BaseController {
                     }
                     if (!isExist && realmStory.getRealmStoryProtos().get(i).getStoryId() != 0) {
                         database.where(RealmStoryProto.class).equalTo("isForReply", false).equalTo("status", MessageObject.STATUS_SENT).equalTo("storyId", realmStory.getRealmStoryProtos().get(i).getStoryId()).findAll().deleteAllFromRealm();
-                        RealmStory userStory = database.where(RealmStory.class).equalTo("userId", userId).findFirst();
+                        RealmStory userStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findFirst();
                         if (userStory != null && userStory.getRealmStoryProtos().size() == 0) {
                             userStory.deleteFromRealm();
                         }
@@ -1242,7 +1242,7 @@ public class MessageDataStorage extends BaseController {
                 database.beginTransaction();
                 RealmRegisteredInfo realmRegisteredInfo = database.where(RealmRegisteredInfo.class).equalTo("id", stories.get(0).getUserId()).findFirst();
                 if (realmRegisteredInfo != null && realmRegisteredInfo.getDisplayName() != null) {
-                    RealmStory realmStory = database.where(RealmStory.class).equalTo("userId", stories.get(0).getUserId()).findFirst();
+                    RealmStory realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", stories.get(0).getUserId()).findFirst();
                     if (realmStory == null) {
                         realmStory = database.createObject(RealmStory.class, SUID.id().get());
                     }
@@ -1314,7 +1314,7 @@ public class MessageDataStorage extends BaseController {
                 if (realmStoryProto != null) {
                     realmStoryProto.deleteFromRealm();
                 }
-                RealmStory userStory = database.where(RealmStory.class).equalTo("userId", userId).findFirst();
+                RealmStory userStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findFirst();
                 if (userStory.getRealmStoryProtos().size() == 0) {
                     userStory.deleteFromRealm();
                 } else {
@@ -1441,7 +1441,7 @@ public class MessageDataStorage extends BaseController {
                 }
 
 
-                RealmStory realmStory = database.where(RealmStory.class).equalTo("userId", storyOwnerUserId).findFirst();
+                RealmStory realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", storyOwnerUserId).findFirst();
                 if (realmStory != null) {
                     int counter = 0;
                     for (int i = 0; i < realmStory.getRealmStoryProtos().size(); i++) {
@@ -1451,7 +1451,10 @@ public class MessageDataStorage extends BaseController {
                     }
 
                     if (counter == realmStory.getRealmStoryProtos().size()) {
-                        database.where(RealmStory.class).equalTo("userId", storyOwnerUserId).findFirst().setSeenAll(true);
+                        RealmStory userRealmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", storyOwnerUserId).findFirst();
+                        if (userRealmStory != null) {
+                            userRealmStory.setSeenAll(true);
+                        }
                     }
 
 
@@ -1548,7 +1551,7 @@ public class MessageDataStorage extends BaseController {
                     realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).notEqualTo("userId", AccountManager.getInstance().getCurrentUser().getId()).sort("lastCreatedAt", Sort.DESCENDING).findAll();
 
                 } else {
-                    realmStory = database.where(RealmStory.class).equalTo("userId", userId).findAll();
+                    realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findAll();
                 }
 
                 if (realmStory != null && realmStory.size() > 0) {
@@ -1607,7 +1610,7 @@ public class MessageDataStorage extends BaseController {
             try {
                 List<StoryObject> storyObjects = new ArrayList<>();
                 MainStoryObject mainStoryObject;
-                RealmStory realmStory = database.where(RealmStory.class).equalTo("userId", userId).findFirst();
+                RealmStory realmStory = database.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findFirst();
                 if (realmStory != null) {
                     if (needSort) {
                         RealmResults<RealmStoryProto> realmStoryProtos = realmStory.getRealmStoryProtos().sort(new String[]{"createdAt", "index"}, new Sort[]{Sort.DESCENDING, Sort.DESCENDING});
@@ -1831,7 +1834,10 @@ public class MessageDataStorage extends BaseController {
         storageQueue.postRunnable(() -> {
             try {
                 database.executeTransaction(realm -> {
-                    realm.where(RealmStory.class).equalTo("userId", userId).findFirst().setSentAll(status);
+                    RealmStory realmStory = realm.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findFirst();
+                    if (realmStory != null) {
+                        realmStory.setSentAll(status);
+                    }
                 });
                 countDownLatch.countDown();
             } catch (Exception e) {
@@ -1913,7 +1919,7 @@ public class MessageDataStorage extends BaseController {
             try {
 
                 database.executeTransaction(realm -> {
-                    RealmStory realmStory = realm.where(RealmStory.class).equalTo("userId", userId).findFirst();
+                    RealmStory realmStory = realm.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findFirst();
                     if (realmStory != null) {
                         realmStory.setSeenAll(seen);
                     }
@@ -1943,7 +1949,7 @@ public class MessageDataStorage extends BaseController {
             try {
 
                 database.executeTransaction(realm -> {
-                    RealmStory realmStory = realm.where(RealmStory.class).equalTo("userId", userId).findFirst();
+                    RealmStory realmStory = realm.where(RealmStory.class).equalTo("sessionId", AccountManager.getInstance().getCurrentUser().getId()).equalTo("userId", userId).findFirst();
                     if (realmStory != null) {
                         realmStory.setIndexOfSeen(position);
                     }
